requirements.txt
flask
paho-mqtt
joblib
scikit-learn
numpy
pandas